fix(ci): satisfy structural gates for connector delivery refactor
